Decimal may be either integer or float
Number.parseFloat parses both integers as well as floats.

Review with @erikap

+++ helpers/mu/sparql.js
@@ -108,7 +108,7 @@ function sparqlEscapeUri( value ){
 };
 
 function sparqlEscapeDecimal( value ){
-  return '"' + Number.parseInt(value) + '"^^xsd:decimal';
+  return '"' + Number.parseFloat(value) + '"^^xsd:decimal';
 };
 
 function sparqlEscapeInt( value ){
